TECH-3032: Split the generic scraper base and custom one into their own files and make the generic one able to run using a --set=start_url=whatever CLI argument
